fix bugs in refactor
