+++ app/src/androidTest/java/com/github/orkest/View/feed/CommentActivityTest.kt
@@ -6,6 +6,7 @@ import androidx.compose.ui.test.*
 import androidx.compose.ui.test.junit4.createComposeRule
 import com.github.orkest.data.Constants
 import com.github.orkest.ViewModel.feed.MockPostViewModel
+import com.github.orkest.ui.feed.CommentActivity
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
